- Live Stream scraping: Added scraping so we don't need to call the api for livestreams all the time. The video queue seems to be reset when the cache needs to be updated. Should probably fix eventually, but ends up with less api calls overall, and I doubt anyone will ever really notice. I also forget if it was in the last commit, but earlier versions didn't page the cache/api correctly. This is now fixed.
- Modularization
Moved config stuff to a separate definition, can be moved to a separate file.

- Parallelization
Fixed browser loads so pages should load in parallel

- Browser optimization
Fixed options so that they're properly passed to the browser

- Refactoring
Moved some stuff around, combined commands when applicable. And changed some debug output to be more understandable.

- Video Queue
Fixed priority queue issues, and made it more in line with what we need the queue to do. (always have live videos at the start and have a comprehensive collection of backlogged content to play)